Move connected event to post-flow-join
The `connected` event was being emitted after the Flowdock connection was
*initiated*, but that meant the event was dispatched before there was actually
a live connection to work with. This meant sending messages on connect, for
example, wouldn't work.

The `connected` event is now correctly dispatched once the connection to
Flowdock is fully established and authed, and all flows have been joined.